Extracted classes into separate files
